# API-Documentation.yml
openapi: 3.0.0
info:
  title: Sensor Service API
  description: API documentation for the Sensor Service in the Microservices HA project.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/sensor:
    post:
      summary: Create a new sensor
      operationId: createSensor
      requestBody:
        description: Sensor details to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
      responses:
        '201':
          description: Sensor created successfully

    get:
      summary: Get all sensors
      operationId: getAllSensors
      responses:
        '200':
          description: A list of sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorResponse'

  /api/sensor/{id}:
    get:
      summary: Get sensor by ID
      operationId: getSensorById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the sensor to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sensor retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'

    put:
      summary: Update sensor by ID
      operationId: updateSensor
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the sensor to update
          schema:
            type: integer
            format: int64
      requestBody:
        description: Sensor details to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
      responses:
        '200':
          description: Sensor updated successfully

    delete:
      summary: Delete sensor by ID
      operationId: deleteSensor
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the sensor to delete
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sensor deleted successfully

components:
  schemas:
    SensorRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        isActive:
          type: boolean
        sensorType:
          $ref: '#/components/schemas/SensorType'

    SensorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        location:
          type: string
        isActive:
          type: boolean
        sensorType:
          $ref: '#/components/schemas/SensorType'

    SensorType:
      type: string
      enum: [OUTDOOR, INDOOR, WATER]