# sensor-service/src/main/java/com/fh/sensorservice/enums/SensorType.java
package com.fh.sensorservice.enums;
/**
 * The SensorType enum represents the types of sensors in the Sensor Service application.
 * It defines the possible categories for sensors, including OUTDOOR, INDOOR, and WATER.
 *
 * This enum is used to categorize sensors based on their intended use or environmental
 * conditions.
 */
public enum SensorType {
    OUTDOOR, INDOOR, WATER
}